streamlit-gui-app/requirements.txt
streamlit>=1.28.0
streamlit-agraph>=0.0.45
plotly>=5.15.0
pandas>=2.0.0
neo4j>=5.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
pydantic>=2.0.0
requests>=2.31.0
websocket-client>=1.6.0